Debugging Shuffle and Search Issues.
Need to keep master "Songs" array otherwise there was an issue when 
clicking the list to be Updated since it truncates the ScrollLists.

Search was not populating the ScrollListFiles which are needed to build the MediaList. 
This was a bug in the UpdateList method that forgot to populate the ScrollListFiles.

When turning off Shuffle, we UpdateFromSelection(0) to make sure the ScrollListFiles is updated. 
Basically if we hit stop first and then unshuffle, the App.SongList.UpdateListFromSelection(0) 
was not getting called and the ScrollListFiles remained shuffled.